refactor(silencer): extract MapDownloader and AmbienceMixer from Game
Move the three async map fetches (UI download, pre-game join fetch,
create-game upload), all map I/O helpers, and the looped background
audio + music selection out of Game into two collaborator classes
that Game owns by value. World is given friend access to both so
they retain Game's prior private-member reach. Behavior is meant
to be unchanged; a suspected regression on the lobby's Create Game
button is flagged in the progress doc for follow-up.
